backend/requirements.txt
# Web crawling and scraping
scrapy>=2.8.0
playwright>=1.25.0
beautifulsoup4>=4.11.1

# API and web framework
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7,<2.0.0  # Keep 1.x for compatibility

# Database
sqlalchemy>=2.0.7
psycopg2-binary>=2.9.5
alembic>=1.10.2
pgvector>=0.1.8

# Vector database
pinecone-client>=2.2.1

# NLP and text processing
spacy>=3.5.1
scikit-learn>=1.2.2
nltk>=3.8.1
transformers>=4.27.2
sentence-transformers>=2.2.2

# Vector similarity and embedding
numpy>=1.24.2
torch>=1.6.0  # Required for sentence-transformers

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pytest>=7.3.1
requests>=2.28.2
python-multipart>=0.0.6

# Monitoring and logging
prometheus-client>=0.16.0
python-json-logger>=2.0.7
